src/journal-app-code/requirements.txt
ulid-py==1.1.0
pytz==2020.4

